+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # 0.1.0
 
  * Created `Add-GitItem` function for promoting new, untracked, and modified files/directories to the Git staging area so they can be committed. Implements the `git add` command.
+ * Created `Get-GitRepository` function for getting an object representing a repository.
 
  
 # 0.0.0 (31 August 2016)